lib: remove undocumented BAREOS_CFGDIR
The configuration parser had set the undocumented and unused environment
variable BAREOS_CFGDIR, which could potentially have been used in
scripts and programs generating configuration snippets. However, as this
was never documented and is not used by Bareos itself, we remove it and
with it the need for setenv() and/or putenv().