python: remove macro MS_WIN32
Python should autodetect 32-bit and 64-bit Windows, but this is broken
when building with GCC[1].
The previous workaround we had in place defined MS_WIN32 or MS_WIN64
depending on what we're building. As MS_WIN32 is always defined, we
should only define MS_WIN64 in case we're building 64-bit Windows.

[1] https://bugs.python.org/issue4709